backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
black==24.2.0
supabase==2.3.4
requests==2.31.0
pytest==8.0.0
pytest-django==4.8.0 